feat: move timeline to separate folder, improve darkmode switch
