+++ src/App.scss
@@ -44,6 +44,3 @@ aside {
 	left: 200px;
 	z-index: 1;
 }
-
-.false * {
-}
