+++ src/components/LightSwitch.tsx
@@ -1,22 +1,20 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState, useEffect } from 'react';
 
-const DarkModeSwitch: React.FC = () => {
+export const DarkModeSwitch: React.FC = () => {
 	const [darkModeOn, setDarkModeOn] = useState<boolean>(
 		localStorage.getItem('darkModeOn')
 			? JSON.parse(localStorage.getItem('darkModeOn')!)
 			: window.matchMedia('(prefers-color-scheme: dark)').matches
-	)
+	);
 
 	useEffect(() => {
-		localStorage.setItem('darkModeOn', JSON.stringify(darkModeOn))
-		console.log('set localStorage')
-	}, [darkModeOn])
+		localStorage.setItem('darkModeOn', JSON.stringify(darkModeOn));
+	}, [darkModeOn]);
 
 	const toggleDarkMode = () => {
-		setDarkModeOn(!darkModeOn)
-		document.documentElement.classList.remove('true', 'false')
-		document.documentElement.classList.add(darkModeOn.toString())
-	}
+		setDarkModeOn(!darkModeOn);
+		document.documentElement.classList.toggle('dark-mode', !darkModeOn);
+	};
 
 	return (
 		<div>
@@ -30,7 +28,7 @@ const DarkModeSwitch: React.FC = () => {
 					<path
 						style={{
 							fill: 'white',
-							filter: 'drop-shadow(3px 5px 2px rgb(0 0 0 / 0.4));',
+							filter: 'drop-shadow(3px 5px 2px rgb(0 0 0 / 0.4));'
 						}}
 						fillRule="evenodd"
 						d="M16.5 6A10.5 10.5 0 0 1 4.7 16.4 8.5 8.5 0 1 0 16.4 4.7l.1 1.3zm-1.7-2a9 9 0 0 1 .2 2 9 9 0 0 1-11 8.8 9.4 9.4 0 0 1-.8-.3c-.4 0-.8.3-.7.7a10 10 0 0 0 .3.8 10 10 0 0 0 9.2 6 10 10 0 0 0 4-19.2 9.7 9.7 0 0 0-.9-.3c-.3-.1-.7.3-.6.7a9 9 0 0 1 .3.8z"
@@ -39,7 +37,7 @@ const DarkModeSwitch: React.FC = () => {
 					<path
 						style={{
 							fill: 'black',
-							filter: 'drop-shadow(3px 5px 2px rgb(255 255 255 / 0.4));',
+							filter: 'drop-shadow(3px 5px 2px rgb(255 255 255 / 0.4));'
 						}}
 						fillRule="evenodd"
 						d="M12 17.5a5.5 5.5 0 1 0 0-11 5.5 5.5 0 0 0 0 11zm0 1.5a7 7 0 1 0 0-14 7 7 0 0 0 0 14zm12-7a.8.8 0 0 1-.8.8h-2.4a.8.8 0 0 1 0-1.6h2.4a.8.8 0 0 1 .8.8zM4 12a.8.8 0 0 1-.8.8H.8a.8.8 0 0 1 0-1.6h2.5a.8.8 0 0 1 .8.8zm16.5-8.5a.8.8 0 0 1 0 1l-1.8 1.8a.8.8 0 0 1-1-1l1.7-1.8a.8.8 0 0 1 1 0zM6.3 17.7a.8.8 0 0 1 0 1l-1.7 1.8a.8.8 0 1 1-1-1l1.7-1.8a.8.8 0 0 1 1 0zM12 0a.8.8 0 0 1 .8.8v2.5a.8.8 0 0 1-1.6 0V.8A.8.8 0 0 1 12 0zm0 20a.8.8 0 0 1 .8.8v2.4a.8.8 0 0 1-1.6 0v-2.4a.8.8 0 0 1 .8-.8zM3.5 3.5a.8.8 0 0 1 1 0l1.8 1.8a.8.8 0 1 1-1 1L3.5 4.6a.8.8 0 0 1 0-1zm14.2 14.2a.8.8 0 0 1 1 0l1.8 1.7a.8.8 0 0 1-1 1l-1.8-1.7a.8.8 0 0 1 0-1z"
@@ -47,6 +45,5 @@ const DarkModeSwitch: React.FC = () => {
 				)}
 			</svg>
 		</div>
-	)
-}
-export default DarkModeSwitch
+	);
+};
